requirements.txt
streamlit>=1.35.0
google-generativeai>=0.8.0
sentence-transformers>=3.0.0
PyPDF2>=3.0.0
numpy>=1.26.0
